Fix segfault where randomx_cache was evicted from under a randomx_vm in light mode.
A randomx_vm in light mode requires the randomx_cache for it's entire lifetime.
A randomx_vm in fast mode requires the randomx_dataset for it's entire lifetime.
This patch resolves bug where in situation like IBD where the cache gets hit frequently,
a required data structure was being evicted and memory released, resulting in segfault.